Remove the npm .bin dir from the PATH
Travis adds that for various reasons, but it leads to us ending up with the wrong
npm version due to the dependency somewhere inside the cordova-android parts. That
npm version is unable to understand the semver tag reference in the gulp-cordova-build-ios
dependency, which leads to rarely weird build failures.

See https://github.com/travis-ci/travis-ci/issues/8813 for some more details.